Renames ratio container to aspect ratio
