+++ website/src/framework/aspect-ratio/aspect-ratio.component.html
@@ -1,5 +1,5 @@
 <div [ngClass]="containerClass">
-    <div class="rc-content">
+    <div class="ar-content">
         <ng-content></ng-content>
     </div>
 </div>
