+++ website/src/framework/content-panel/content-panel.component.html
@@ -1,7 +1,7 @@
 <div [ngClass]="rootNgClass" [hidden]="hidden">
-    <td-ratio-container ratio="3:2">
+    <td-aspect-ratio ratio="3:2">
         <td-illustration [illustration]="panel.illustration" [visible]="!hidden"></td-illustration>
-    </td-ratio-container>
+    </td-aspect-ratio>
     <div *ngIf="contentTextPanel" class="cp-text-area">
         <div>
             <h2 *ngIf="isMobile$ | async" class="text-18-26 title-standalone">
