+++ website/src/framework/content-panel/content-panel.component.scss
@@ -17,7 +17,7 @@
     }
 }
 
-td-ratio-container {
+td-aspect-ratio {
     flex: 2 0 424px;
     max-width: 100%;
