+++ website/src/framework/illustration/captioned-illustration.component.html
@@ -1,13 +1,13 @@
 <div class="section ci-root">
     <div class="ci-illustration">
         <div class="ci-illustration-header"></div>
-        <td-ratio-container ratio="16:9">
+        <td-aspect-ratio ratio="16:9">
             <td-illustration
                 class="il-image-illustration"
                 [illustration]="illustration"
                 [visible]="true"
             ></td-illustration>
-        </td-ratio-container>
+        </td-aspect-ratio>
         <div class="ci-illustration-footer"></div>
     </div>
     <td-rich-text [value]="caption" class="text-24-32 text-20-32-tablet text-14-21-mobile"></td-rich-text>
