+++ website/src/page/webinars-page/webinars-page.component.html
@@ -9,9 +9,9 @@
             <div class="subsection-margin wp-primary-webinar card card-padding">
                 <div *ngIf="!primaryWebinar" class="wp-primary-placeholder"></div>
                 <ng-container *ngIf="primaryWebinar">
-                    <td-ratio-container ratio="16:9">
+                    <td-aspect-ratio ratio="16:9">
                         <img [src]="primaryWebinar.imageURL" class="wp-img wp-img-primary" [alt]="" />
-                    </td-ratio-container>
+                    </td-aspect-ratio>
                     <div class="wp-primary-webinar-summary">
                         <h3 class="wp-webinar-title-container text-24-32 text-22-33-tablet text-18-32-mobile">
                             <span class="wp-webinar-title">{{ primaryWebinar.title | plainText }}</span>
@@ -47,9 +47,9 @@
                 <ul class="wp-webinar-list">
                     <li *ngFor="let webinar of allWebinars" class="wp-webinar-list-item">
                         <div class="wp-img-tertiary-container">
-                            <td-ratio-container ratio="16:9" class="wp-img-tertiary">
+                            <td-aspect-ratio ratio="16:9" class="wp-img-tertiary">
                                 <img [src]="webinar.imageURL" class="wp-img" [alt]="" />
-                            </td-ratio-container>
+                            </td-aspect-ratio>
                             <h3 class="text-18-26 wp-webinar-title hide-gt-mobile">
                                 {{ webinar.title | plainText }}
                             </h3>
