+++ website/src/page/white-papers-page/white-papers-page.component.scss
@@ -18,7 +18,7 @@
         gap: 12px;
     }
 
-    td-ratio-container {
+    td-aspect-ratio {
         flex: 0 0 50%;
 
         @media (max-width: $media-max-width-mobile) {
